si-ks25 pipeline
